requirements.txt
# LangChain OSINT System Requirements
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.1.0

# UI Framework
streamlit>=1.30.0
streamlit-chat>=0.1.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Docker Integration
docker>=6.1.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0

fastapi==0.104.1
uvicorn[standard]==0.24.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
dnspython==2.4.2
python-whois==0.8.0
shodan==1.30.1
censys==2.2.10
python-nmap==0.7.1
selenium==4.16.0
lxml==4.9.3
python-multipart==0.0.6
jinja2==3.1.2 